LLM/gemini-sample/region-test/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
pandas>=2.0.0
plotly>=5.0.0
